make setReduce private for now
